+++ src/constants/screen-size.js
@@ -1,7 +1,13 @@
+const sm = 576
+const md = 768
+const lg = 992
+const xl = 1200
 export default {
-  smallScreenMinWidth: 480,
-  smallScreenMaxWidth: 768,
-  mediumScreenMaxWidth: 992,
-  mediumScreenMinWidth: 769,
-  largeScreenMinWidth: 993
+  smallScreenMinWidth: sm,
+  smallScreenMaxWidth: md - 1,
+  mediumScreenMaxWidth: lg -1,
+  mediumScreenMinWidth: md,
+  largeScreenMinWidth: lg,
+  largeScreenMaxWidth: xl - 1,
+  xLargeScreenMinWidth: xl
 }