+++ src/containers/Article.js
@@ -1,37 +1,37 @@
 /* eslint no-console:0 */
 'use strict'
 
-import * as ArticleComponents from '../components/article/index'
-import ArticleMeta from '../components/article/article-meta'
 import ArticleTools from './ArticleTools'
-import Header from '@twreporter/react-components/lib/header'
 import Helmet from 'react-helmet'
-import LeadingVideo from '../components/shared/LeadingVideo'
 import License from '../components/shared/License'
+import LogoIcon from '../../static/asset/icon-placeholder.svg'
 import PropTypes from 'prop-types'
 import React, { PureComponent } from 'react'
 import ReadingProgress from '../components/article/ReadingProgress'
 import SystemError from '../components/SystemError'
-import TitleRowUpon from '../components/article/title-row-upon'
-import TitleRowAbove from '../components/article/title-row-above'
 import commonStyles from '../components/article/Common.scss'
+import constPropTypes from '../constants/prop-types'
 import cx from 'classnames'
 import deviceConst from '../constants/device'
-import leadingImgStyles from '../components/article/LeadingImage.scss'
-import LogoIcon from '../../static/asset/icon-placeholder.svg'
-import withLayout, { defaultTheme, photoTheme } from '../helpers/with-layout'
+import constPageThemes from '../constants/page-themes'
+import constStyledComponents from '../constants/styled-components'
 import styled from 'styled-components'
 import styles from './Article.scss'
-import { screen } from '../themes/screen'
 import twreporterRedux from '@twreporter/redux'
-import { PHOTOGRAPHY_ARTICLE_STYLE, SITE_META, SITE_NAME, appId } from '../constants/index'
-import { globalColor, colors, componentMargin, layout, letterSpace } from '../themes/common-variables'
+import withLayout from '../helpers/with-layout'
+import layoutMaker from '../components/article/layout/layout-maker'
+import { Body } from '../components/article/Body'
+import { BottomRelateds } from '../components/article/BottomRelateds'
+import { BottomTags } from '../components/article/BottomTags'
+import { Introduction } from '../components/article/Introduction'
+import { PHOTOGRAPHY_ARTICLE_STYLE, SITE_META, SITE_NAME } from '../constants/index'
+import { articleLayout as layout } from '../themes/layout'
 import { camelizeKeys } from 'humps'
 import { connect } from 'react-redux'
-import { getAbsPath, getScreenType } from '../utils/index'
-
-//testing
-import { TITLE_POSITION_ABOVE, TITLE_POSITION_UPON_LEFT,  NAVBAR_POSITION_UPON } from '../constants/page-themes'
+import { date2yyyymmdd } from '../utils/date'
+import { getScreenType } from '../utils/screen'
+import { colors } from '../themes/common-variables'
+import { screen } from '../themes/screen'
 
 // lodash
 import forEach from 'lodash/forEach'
@@ -57,48 +57,36 @@ const _ = {
 const { actions, reduxStateFields, utils } = twreporterRedux
 const { fetchAFullPost } = actions
 
-const ArticleContainer = styled.div`
-  background-color: ${ props => (props.bgColor ? props.bgColor : colors.gray.lightGray) };
-  min-height: 20em;
-  padding-top: ${ props => (props.titlePosition === TITLE_POSITION_UPON_LEFT ? '0' : '20px') };
-`
-
-const Content = styled.div`
-  color: ${ props => (props.fontColor ? props.fontColor : globalColor.textColor)};
+const ArticleContainer = styled.article`
+  color: ${props => props.fontColor ? props.fontColor : colors.gray.gray25};
   a {
-      border-bottom: 1px ${globalColor.primaryColor} solid;
+      border-bottom: 1px ${colors.primaryColor} solid;
       cursor: pointer;
       transition: 0.5s color ease;
       position: relative;
-      color: ${ props => (props.fontColor ? props.fontColor : globalColor.textColor)};
-      letter-spacing: ${letterSpace.generalLetterSpace};
+      color: ${ props => (props.fontColor ? props.fontColor : colors.gray.gray25)};
       &:hover {
-        color: ${globalColor.primaryColor};
+        color: ${colors.primaryColor};
       }
+  }
+
+  u {
+    text-decoration: none;
   }
 `
 
 const IntroductionContainer = styled.div`
   display: block;
-  margin: 0 auto ${componentMargin.doubleMarginBottom} auto;
+  margin: 0 auto 80px auto;
   ${screen.desktopAbove`
-    width: ${layout.desktop.small};
-  `}
+    width: ${layout.desktop.width.small}px;
+  `};
   ${screen.tablet`
-    width: ${layout.tablet.small};
-  `}
+    width: ${layout.tablet.width.small}px;
+  `};
   ${screen.mobile`
-    margin: 0 ${componentMargin.horizontalMargin} ${componentMargin.doubleMarginBottom} ${componentMargin.horizontalMargin};
-  `}
-`
-
-const HeaderContainer = styled.div`
-  position: absolute;
-  top: 0;
-  width: 100%;
-  @media (max-width: 1023px) {
-    display: none;
-  }
+    margin: 0 24px 80px 24px;
+  `};
 `
 
 const scrollPosition = {
@@ -108,23 +96,25 @@ const scrollPosition = {
 
 const ArticlePlaceholder = () => {
   return (
-    <div className={cx(styles['placeholder'])}>
-      <div className={cx(styles['title-row'], commonStyles['inner-block'])}>
-        <div className={styles['ph-title-1']}></div>
-        <div className={styles['ph-title-2']}></div>
-        <div className={styles['ph-author']}></div>
-      </div>
-      <div className={cx(styles['leading-img'], leadingImgStyles['leading-img'])}>
-        <div className={styles['ph-image']}>
-          <LogoIcon className={styles['logo-icon']} />
+    <constStyledComponents.ResponsiveContainerForAritclePage>
+      <div className={cx(styles['placeholder'])}>
+        <div className={cx(styles['title-row'], commonStyles['inner-block'])}>
+          <div className={styles['ph-title-1']}></div>
+          <div className={styles['ph-title-2']}></div>
+          <div className={styles['ph-author']}></div>
         </div>
+        <div className={styles['leading-img']}>
+          <div className={styles['ph-image']}>
+            <LogoIcon className={styles['logo-icon']} />
+          </div>
+        </div>
+        <IntroductionContainer>
+          <div className={styles['ph-content']}></div>
+          <div className={styles['ph-content']}></div>
+          <div className={styles['ph-content-last']}></div>
+        </IntroductionContainer>
       </div>
-      <IntroductionContainer>
-        <div className={styles['ph-content']}></div>
-        <div className={styles['ph-content']}></div>
-        <div className={styles['ph-content-last']}></div>
-      </IntroductionContainer>
-    </div>
+    </constStyledComponents.ResponsiveContainerForAritclePage>
   )
 }
 
@@ -151,6 +141,7 @@ class Article extends PureComponent {
 
     this._onScroll = _.throttle(this._handleScroll, 300).bind(this)
     this._handleScroll = this._handleScroll.bind(this)
+    this.changeFontSize = this._changeFontSize.bind(this)
 
     // reading progress component
     this.rp = null
@@ -171,7 +162,7 @@ class Article extends PureComponent {
     }
   }
 
-  changeFontSize(fontSize) {
+  _changeFontSize(fontSize) {
     this.setState({
       fontSize:fontSize,
       isFontSizeSet:true
@@ -324,24 +315,9 @@ class Article extends PureComponent {
     }
   }
 
-  _composeAuthors(article) {
-    article = article || {}
-    let authors = []
-    const list = [ 'writters', 'photographers', 'designers', 'engineers' ]
-    list.forEach((item) => {
-      if (Array.isArray(article[item])) {
-        article[item].forEach((author) => {
-          // remove 's'. writters -> writter
-          let type = item.slice(0, -1)
-          authors.push(_.merge({}, author, { type: type }))
-        })
-      }
-    })
-    return authors
-  }
-
   render() {
     const { entities, params, selectedPost, theme } = this.props
+    const { fontSize } = this.state
     const error = _.get(selectedPost, 'error')
     if (error) {
       return (
@@ -360,55 +336,35 @@ class Article extends PureComponent {
     const slug = _.get(selectedPost, 'slug', '')
     const article = camelizeKeys(_.get(postEntities, slug))
     const articleStyle = _.get(article, 'style')
-    // const outerClass = (articleStyle===PHOTOGRAPHY_ARTICLE_STYLE) ?
-                //  cx(styles['article-container'], styles['photo-container']) : styles['article-container']
-    const contentClass = (articleStyle===PHOTOGRAPHY_ARTICLE_STYLE) ?
-                 cx(styles['article-inner'], styles['photo-page-inner']) : styles['article-inner']
     const isFetching = _.get(selectedPost, 'isFetching')
 
     const relateds = camelizeKeys(utils.denormalizePosts(_.get(article, 'relateds', []), postEntities))
     const topics = camelizeKeys(utils.denormalizeTopics(_.get(article, 'topics', []), topicEntities, postEntities))
     const topic = topics[0]
-    const authors = this._composeAuthors(article)
     const bodyData = _.get(article, [ 'content', 'apiData' ], [])
-    const leadingVideo = _.get(article, 'leadingVideo', null)
-    const heroImage = _.get(article, [ 'heroImage' ], null)
-    const heroImageSize = _.get(article, [ 'heroImageSize' ], 'normal')
     const introData = _.get(article, [ 'brief', 'apiData' ], [])
-    const cUrl = getAbsPath(this.context.location.pathname, this.context.location.search)
-
     const topicName = _.get(topic, 'topicName')
     const topicArr = _.get(topic, 'relateds')
 
-    // const updatedAt = _.get(article, 'updatedAt') || _.get(article, 'publishedDate')
-
     // for head tag
     const canonical = SITE_META.URL + 'a/' + slug
-    const articleTitle = _.get(article, 'title', '')
-    const ogTitle = articleTitle + SITE_NAME.SEPARATOR + SITE_NAME.FULL
+    const ogTitle = _.get(article, 'title', '') + SITE_NAME.SEPARATOR + SITE_NAME.FULL
     const ogDesc = _.get(article, 'ogDescription', SITE_META.DESC)
     const ogImage = _.get(article, 'ogImage.resizedTargets.mobile.url', SITE_META.OG_IMAGE)
 
-    const pathname = _.get(this.props, 'location.pathname')
-
-    // theme
-    const _theme = camelizeKeys(theme)
-    const bgColor = _.get(_theme, 'bgColor')
-    // const footerBgColor = _.get(_theme, 'footer_bg_color')
-    const fontColor = _.get(_theme, 'fontColor')
-    const titlePosition = _.get(_theme, 'titlePosition')
-    const headerPosition = _.get(_theme, 'headerPosition')
-    const titleColor = _.get(_theme, 'titleColor')
-    const subTitleColor = _.get(_theme, 'subtitleColor')
-    const topicColor = _.get(_theme, 'topicColor')
-    const logoColor =  _.get(_theme, 'logoColor')
-    const fontColorSet = {
-      topicFontColor: topicColor,
-      titleFontColor: titleColor,
-      subtitleFontColor: subTitleColor
-    }
-
     const license = _.get(article, 'copyright', 'Creative-Commons')
+
+
+    // render leading components, including
+    // title, subtitle, topic name, hero image, leading video, authors and share buttons.
+    // Those components will be rendered in different orders on demand
+    const layoutJSX = layoutMaker.renderLayout({
+      article,
+      topic,
+      theme,
+      fontSize,
+      changeFontSize: this.changeFontSize
+    })
 
     return (
       <div>
@@ -432,133 +388,54 @@ class Article extends PureComponent {
           ]}
         />
         <div itemScope itemType="http://schema.org/Article">
-          {isFetching ? <ArticleContainer bgColor="transparent"><ArticlePlaceholder /></ArticleContainer> :
-          <ArticleContainer
-            bgColor={bgColor}
-            titlePosition={titlePosition}
-          >
-            {
-              leadingVideo ?
-                <LeadingVideo
-                  filetype={_.get(leadingVideo, 'filetype')}
-                  title={_.get(leadingVideo, 'title')}
-                  src={_.get(leadingVideo, 'url')}
-                  poster={_.get(heroImage, 'resizedTargets')}
-                /> : null
-            }
-            <ReadingProgress ref={ele => this.rp = ele}/>
-            <Content
-              fontColor={fontColor}
-            >
-              <article ref={div => {this.progressBegin = div}} className={contentClass}>
-                {
-                  titlePosition === TITLE_POSITION_ABOVE ?
-                  <div>
-                    <TitleRowAbove
-                      article={article}
-                      topic={topic}
-                      canonical={canonical}
-                      fontColorSet={fontColorSet}
-                    />
-                    <ArticleMeta
-                      authors={authors}
-                      extendByline={_.get(article, 'extendByline')}
-                      publishedDate={article.publishedDate}
-                      appId={appId}
-                      url={cUrl}
-                      title={article.title}
-                      changeFontSize={(fontSize)=>this.changeFontSize(fontSize)}
-                      fontSize={this.state.fontSize}
+          {isFetching ? <ArticlePlaceholder /> :
+              <React.Fragment>
+                <ReadingProgress ref={ele => this.rp = ele}/>
+                <ArticleContainer
+                  fontColor={theme.color.font}
+                  ref={div => {this.progressBegin = div}}
+                >
+                  <div itemProp="publisher" itemScope itemType="http://schema.org/Organization">
+                    <meta itemProp="name" content="報導者" />
+                    <meta itemProp="email" content="contact@twreporter.org" />
+                    <link itemProp="logo" href="https://www.twreporter.org/asset/logo-large.png" />
+                    <link itemProp="url" href="https://www.twreporter.org/" />
+                  </div>
+                  <link itemProp="mainEntityOfPage" href={canonical} />
+                  <meta itemProp="dateModified" content={date2yyyymmdd(_.get(article, 'updatedAt'))} />
+                  {layoutJSX}
+                  <div
+                    id="article-body"
+                    ref={node => this.articleBody = node}
+                  >
+                    <IntroductionContainer>
+                      <Introduction
+                        data={introData}
+                        fontSize={fontSize}
+                      />
+                    </IntroductionContainer>
+                    <Body
+                      data={bodyData}
+                      fontSize={fontSize}
+                      articleStyle={articleStyle}
                     />
                   </div>
-                : null
-                }
-
-                {
-                  !leadingVideo && !this.props.ifDelegateImage ?
-                    <div className={styles['leading-img']}>
-                      <ArticleComponents.LeadingImage
-                        size={heroImageSize}
-                        image={_.get(heroImage, 'resizedTargets')}
-                        id={_.get(heroImage, 'id')}
-                        description={_.get(article, 'leadingImageDescription', '')}
-                        titlePosition={titlePosition}
-                      >
-                        {
-                          headerPosition === NAVBAR_POSITION_UPON ?
-                            <HeaderContainer>
-                              <Header
-                                isIndex={false}
-                                pathName={pathname}
-                                fontColor={fontColor}
-                                bgColor={''}
-                                logoColor={logoColor}
-                                headerPosition={headerPosition}
-                              />
-                            </HeaderContainer>
-                          : null
-                        }
-                        {
-                          titlePosition === TITLE_POSITION_UPON_LEFT ?
-                            <TitleRowUpon
-                              article={article}
-                              topic={topic}
-                              canonical={canonical}
-                              fontColorSet={fontColorSet}
-                            />
-                          : null
-                        }
-                      </ArticleComponents.LeadingImage>
-                    </div> : null
-                }
-
-                {
-                  titlePosition === TITLE_POSITION_UPON_LEFT ?
-                    <ArticleMeta
-                      authors={authors}
-                      extendByline={_.get(article, 'extendByline')}
-                      publishedDate={article.publishedDate}
-                      appId={appId}
-                      url={cUrl}
-                      title={article.title}
-                      changeFontSize={(fontSize)=>this.changeFontSize(fontSize)}
-                      fontSize={this.state.fontSize}
-                    />
-                  : null
-                }
-                <div
-                  id="article-body"
-                  ref={node => this.articleBody = node}
+                </ArticleContainer>
+                <License license={license} publishedDate={article.publishedDate}/>
+                <constStyledComponents.ResponsiveContainerForAritclePage
+                  size="small"
                 >
-                  <IntroductionContainer>
-                    <ArticleComponents.Introduction
-                      data={introData}
-                      fontSize={this.state.fontSize}
-                    />
-                  </IntroductionContainer>
-                  <ArticleComponents.Body
-                    data={bodyData}
-                    fontSize={this.state.fontSize}
-                    articleStyle={articleStyle}
+                  <BottomTags
+                    data={article.tags}
                   />
-                </div>
-              </article>
-            </Content>
-            <License license={license} publishedDate={article.publishedDate}/>
-            <div className={cx(commonStyles['components'], 'hidden-print', styles['padding-patch'])}>
-              <div className={cx('inner-max', commonStyles['component'])}>
-                <ArticleComponents.BottomTags
-                  data={article.tags}
-                />
-              </div>
-              <ArticleComponents.BottomRelateds
-                relateds={relateds}
-                currentId={article.id}
-                topicName={topicName}
-                topicArr={topicArr}
-              />
-            </div>
-          </ArticleContainer>
+                  <BottomRelateds
+                    relateds={relateds}
+                    currentId={article.id}
+                    topicName={topicName}
+                    topicArr={topicArr}
+                  />
+                </constStyledComponents.ResponsiveContainerForAritclePage>
+              </React.Fragment>
           }
           <div className="hidden-print">
             <ArticleTools
@@ -576,12 +453,33 @@ export function mapStateToProps(state) {
   const selectedPost = state[reduxStateFields.selectedPost]
   const post = _.get(entities, [ reduxStateFields.postsInEntities, selectedPost.slug ], {})
   const style = post.style
-  let theme = defaultTheme
-  // backwards compatible for photo articles
-  if (style === PHOTOGRAPHY_ARTICLE_STYLE) {
-    theme = photoTheme
+  let theme
+
+  if (!_.get(post, 'theme')) {
+    // backwards compatible for photo articles
+    if (style === PHOTOGRAPHY_ARTICLE_STYLE) {
+      theme = constPageThemes.photoTheme
+    } else {
+      theme = constPageThemes.defaultTheme
+    }
+  } else {
+    const rawTheme = post.theme
+    theme = {
+      color: {
+        bg: rawTheme.bg_color,
+        font: rawTheme.font_color,
+        footerBg: rawTheme.footer_bg_color,
+        logo: rawTheme.logo_color,
+        title: rawTheme.title_color,
+        subtitle: rawTheme.subtitle_color,
+        topic: rawTheme.topic_color
+      },
+      position: {
+        header: rawTheme.header_position,
+        title: rawTheme.title_position
+      }
+    }
   }
-  theme = post.theme || theme
 
   return {
     entities,
@@ -600,18 +498,16 @@ Article.contextTypes = {
 
 Article.propTypes = {
   entities: PropTypes.object,
-  ifDelegateImage: PropTypes.bool,
   params: PropTypes.object,
   selectedPost: PropTypes.object,
-  theme: PropTypes.object
+  theme: constPropTypes.theme
 }
 
 Article.defaultProps = {
   entities: {},
-  ifDelegateImage: false,
   params: {},
   selectedPost: {},
-  theme: defaultTheme
+  theme: constPageThemes.defaultTheme
 }
 
 export { Article }