+++ src/containers/Topics.js
@@ -10,11 +10,12 @@ import styled from 'styled-components'
 import withLayout from '../helpers/with-layout'
 import twreporterRedux from '@twreporter/redux'
 import uniq from 'lodash/uniq'
-import { BRIGHT, LINK_PREFIX, SITE_META, SITE_NAME } from '../constants/'
+import { LINK_PREFIX, SITE_META, SITE_NAME } from '../constants/'
 import { InternalServerError } from '../custom-error'
 import { TopicsList } from '@twreporter/react-components/lib/listing-page'
 import { connect } from 'react-redux'
-import { date2yyyymmdd, formatPostLinkTarget, formatPostLinkTo } from '../utils'
+import { date2yyyymmdd } from '../utils/date'
+import { formatPostLinkTarget, formatPostLinkTo } from '../utils/url'
 
 const _ = {
   concat,
@@ -211,8 +212,6 @@ function mapStateToProps(state, ownProps) {
   const topicListError = _.get(topicList, 'error', null)
   const topicError = _.get(selectedTopic, 'error', null)
 
-  const pageTheme = _.get(state, 'header.pageTheme', BRIGHT)
-
   return ({
     page: locationPage,
     nPerPage,
@@ -222,8 +221,7 @@ function mapStateToProps(state, ownProps) {
     isTopicsFetching,
     topicListError,
     topicError,
-    pathname,
-    pageTheme
+    pathname
   })
 }